fix: keep the classifier core out of the boot bundle
classifierCues is eager (main.ts imports it), and it imported
classifierParamsForFrame as a value from deriveClassification, which pulled the
whole classifier algorithm (spatial hash, verticality cue, morphology) into the
index chunk and pushed it 8 KiB over its ceiling. Move the preset numbers and
the unit-scaling helper to a light classifierFrame leaf that holds only data and
a pure function. deriveClassification imports the numbers back, so there is one
source of truth and the preset digest is unchanged.
